fix: recompile glibc
This diff adds a new feature by which we refuse to install
glibc by default unless a file is present on the system.

We will add this file when compiling the next version of LFS
and otherwise we won't be able to reinstall glibc.

Additionally, this diff includes new patches that fix the
following security vulnerabilities:

- https://www.linuxfromscratch.org/blfs/advisories/consolidated.html#sa-12.0-018
- https://www.linuxfromscratch.org/blfs/advisories/consolidated.html#sa-12.0-005
- https://www.linuxfromscratch.org/blfs/advisories/consolidated.html#sa-12.0-004

As recommended by the book, I have rebuilt glibc and replaced
the dynamic linker as well as libc.a and libc.so.6.