fix: attempt to track all the installed files (#1)
This diff includes work to track all installed files. While there, we
also simplify how we handle configuration files and we make sure we
strip debug symbols from binary files, which leads to a smaller system.

We're not there yet in terms of tracking all files and the main
offenders are:

* `/etc/hosts.new`
* `/etc/ld.so.cache.new`
* `./usr/lib/libstdc++.so.6.0.32-gdb.py`

Apart from that, we know the origin of all other files installed on the
system or the files presented on the system have been generated by other
programs and it is safe to ignore them (e.g., the manifest files
generated by packages).